+++ SemesterProject/Server/Models/UserStats.cs
@@ -9,12 +9,12 @@ namespace SemesterProject.Server.Models
     internal class UserStats
     {
         public string UserName { get; }
-        public int Elo { get; }
+        public double Elo { get; }
         public int Wins { get; }
         public int Loses { get; }
         public double WinLoseRatio { get; }
 
-        public UserStats(string userName, int elo, int wins, int loses, double winRate)
+        public UserStats(string userName, double elo, int wins, int loses, double winRate)
         {
             UserName = userName;
             Elo = elo;
